Add task to compile Elm modules
The result of the compilation does not get imported by `application.js`.
By placing it directly in the `public` directory, it does not get
wrapped by module loading code. It has to be referenced by a separate
<script> tag and does not get minified, though.

This task will be removed once asset management has been migrated to
Webpack which is planned for the future.